Changed nick of annoucer
A bug appeared in the current 2.0 version of m_chanlog.so: Since the m_chanlog.so makes annoucements using the server name as "nick", and this "nick" isn't officially online on the server, this causes bot software to crash since it tries to look up a nick on network which does not exist.

This solves the problem by pretending to be ChanServ when sending notices to the channel, so the bots which expect an user that writes in the channel to also be online on the server aswell, does not crash.